split up large flyout_menu component
