+++ packages/eui/src/components/flyout/flyout_menu/index.tsx
@@ -1,0 +1,17 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+export type {
+  EuiFlyoutMenuBackButtonProps,
+  EuiFlyoutHistoryItem,
+  EuiFlyoutMenuPagination,
+  EuiFlyoutMenuCustomAction,
+  EuiFlyoutMenuAction,
+  EuiFlyoutMenuProps,
+} from './types';
+export { EuiFlyoutMenu } from './flyout_menu';
